% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pkg_report.R
\name{create_pkg_report}
\alias{create_pkg_report}
\title{Generate a package report in HTML}
\usage{
create_pkg_report(pkgname, path = paste(pkgname, "- summary report"))
}
\arguments{
\item{pkgname}{String containing the name of the package.}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file
extension}. For example, \code{"my package report"}.}
}
\description{
Create a report in HTML
}
