% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_argument.R
\name{extract_argument}
\alias{extract_argument}
\title{Extract the arguments of a function for a loaded package}
\usage{
extract_argument(fnc_name, package_name)
}
\arguments{
\item{fnc_name}{String containing the name of the function.}

\item{package_name}{String containing the name of the loaded package.}
}
\description{
Returns a character vector containing the arguments of a
function for a loaded package. Arguments are separated by semi-colons.
}
\examples{
extract_argument(
  fnc_name = "extract_argument",
  package_name = "devtoolbox")

}
