% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_family.R
\name{extract_family}
\alias{extract_family}
\title{Extract the family of a function for a loaded package}
\usage{
extract_family(fnc_name, package_name)
}
\arguments{
\item{fnc_name}{String containing the name of the function.}

\item{package_name}{String containing the name of the loaded package.}
}
\description{
Returns a character vector containing the family of a
function for a loaded package.
}
\examples{
extract_family(
  fnc_name = "extract_family",
  package_name = "devtoolbox")

}
