% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_title.R
\name{extract_title}
\alias{extract_title}
\title{Extract the title of a function for a loaded package}
\usage{
extract_title(fnc_name, package_name)
}
\arguments{
\item{fnc_name}{String containing the name of the function.}

\item{package_name}{String containing the name of the loaded package.}
}
\description{
Returns a character vector containing the title of a
function for a loaded package.
}
\examples{
extract_title(
  fnc_name = "extract_title",
  package_name = "devtoolbox")

}
