% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary.R
\name{generate_summary}
\alias{generate_summary}
\title{Run a summary of functions and the associated description for a loaded
package}
\usage{
generate_summary(package_name)
}
\arguments{
\item{package_name}{String providing the name of the loaded package, e.g. surveytoolbox}
}
\value{
A data frame summarising functions, family, title, description, and
arguments of a loaded package.
}
\description{
Return a data frame summarising functions, family, title,
description, and arguments of a loaded package.
}
\examples{
library(devtoolbox)
generate_summary(package_name = "devtoolbox")

}
