% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gh_pr.R
\name{get_gh_pr}
\alias{get_gh_pr}
\title{Get pull request information from GitHub for the given time period.}
\usage{
get_gh_pr(owner, repo, start_date, end_date)
}
\arguments{
\item{owner}{String specifying the owner of the repository.}

\item{repo}{String specifying the name of the repo.}

\item{start_date}{String specifying start date of reporting period to filter
by, in the format \code{YYYY-MM-DD}.}

\item{end_date}{String specifying end date of reporting period to filter by,
in the format \code{YYYY-MM-DD}.}
}
\description{
Get information on pull request from the specified GitHub
repository. This is a wrapper around the 'gh' package. Original code taken
from \url{https://github.com/jennybc/analyze-github-stuff-with-r/}.
}
\examples{
get_gh_pr(owner = "martinctc", repo = "rwa")

}
