% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help_file.R
\name{get_help_file}
\alias{get_help_file}
\title{Extract help file function based on 'utils:::.getHelpFile'}
\usage{
get_help_file(file)
}
\arguments{
\item{file}{String containing name of file.}
}
\description{
Extract help file function based on 'utils:::.getHelpFile'
}
