% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_dependencies.R
\name{recursive_dependencies}
\alias{recursive_dependencies}
\title{Compute recursive dependencies of packages}
\usage{
recursive_dependencies(package, db, seen_packages = NULL)
}
\arguments{
\item{package}{String specifying name of original package.}

\item{db}{Parameter to pass to 'tools::package_dependencies'}

\item{seen_packages}{String containing names of packages to exclude}
}
\description{
Compute recursive package dependencies of packages. Credits to 'pkgnet' for
the original function.
}
